day2/python_base/python_package_index/requirements.txt
requests>=2.26.0
python-dotenv>=0.19.0
pytest>=7.0.0
black>=22.0.0
mypy>=0.900 